+++ src/CSharpLanguageServer/Handlers/Initialization.fs
@@ -61,6 +61,12 @@ module Initialization =
 
         MSBuildLocator.RegisterInstance(vsInstance)
 
+(*
+        logger.trace (
+            Log.setMessage "handleInitialize: p.Capabilities={caps}"
+            >> Log.addContext "caps" (serialize p.Capabilities)
+        )
+*)
         scope.Emit(ClientCapabilityChange p.Capabilities)
 
         // TODO use p.RootUri
@@ -80,6 +86,7 @@ module Initialization =
 
     let handleInitialized (lspClient: ILspClient)
                           (stateActor: MailboxProcessor<ServerStateEvent>)
+                          (getRegistrations: (ClientCapabilities option) -> Registration list)
                           (scope: ServerRequestScope)
                           (_p: InitializedParams)
             : Async<LspResult<unit>> =
@@ -88,46 +95,23 @@ module Initialization =
                 Log.setMessage "handleInitialized: \"initialized\" notification received from client"
             )
 
-            //
-            // registering w/client for didChangeWatchedFiles notifications"
-            //
-            let clientSupportsWorkspaceDidChangeWatchedFilesDynamicReg =
-                scope.ClientCapabilities
-                |> Option.bind (fun x -> x.Workspace)
-                |> Option.bind (fun x -> x.DidChangeWatchedFiles)
-                |> Option.bind (fun x -> x.DynamicRegistration)
-                |> Option.defaultValue true
+            let registrationParams = { Registrations = getRegistrations scope.ClientCapabilities |> List.toArray }
 
-            match clientSupportsWorkspaceDidChangeWatchedFilesDynamicReg with
-            | true ->
-                let fileChangeWatcher = { GlobPattern = U2.First "**/*.{cs,csproj,sln}"
-                                          Kind = None }
-
-                let didChangeWatchedFilesRegistration: Registration =
-                    { Id = "id:workspace/didChangeWatchedFiles"
-                      Method = "workspace/didChangeWatchedFiles"
-                      RegisterOptions = { Watchers = [| fileChangeWatcher |] } |> serialize |> Some
-                    }
-
-                try
-                    let! regResult =
-                        lspClient.ClientRegisterCapability(
-                            { Registrations = [| didChangeWatchedFilesRegistration |] })
-
-                    match regResult with
-                    | Ok _ -> ()
-                    | Error error ->
-                        logger.warn(
-                            Log.setMessage "handleInitialized: didChangeWatchedFiles registration has failed with {error}"
-                            >> Log.addContext "error" (string error)
-                        )
-                with
-                | ex ->
+            // TODO: Retry on error?
+            try
+                match! lspClient.ClientRegisterCapability registrationParams with
+                | Ok _ -> ()
+                | Error error ->
                     logger.warn(
                         Log.setMessage "handleInitialized: didChangeWatchedFiles registration has failed with {error}"
-                        >> Log.addContext "error" (string ex)
+                        >> Log.addContext "error" (string error)
                     )
-            | false -> ()
+            with
+            | ex ->
+                logger.warn(
+                    Log.setMessage "handleInitialized: didChangeWatchedFiles registration has failed with {error}"
+                    >> Log.addContext "error" (string ex)
+                )
 
             //
             // retrieve csharp settings