+++ src/CSharpLanguageServer/Handlers/SelectionRange.fs
@@ -1,0 +1,14 @@
+namespace CSharpLanguageServer.Handlers
+
+open Ionide.LanguageServerProtocol.Types
+
+open CSharpLanguageServer.State
+
+[<RequireQualifiedAccess>]
+module SelectionRange =
+    let provider (clientCapabilities: ClientCapabilities option) : bool option = None
+
+    let registration (clientCapabilities: ClientCapabilities option) : Registration option = None
+
+    let handle (scope: ServerRequestScope) (p: SelectionRangeParams) : AsyncLspResult<SelectionRange list option> =
+        LspResult.notImplemented<SelectionRange list option> |> async.Return