+++ src/CSharpLanguageServer/Handlers/SemanticTokens.fs
@@ -3,15 +3,17 @@ namespace CSharpLanguageServer.Handlers
 open System
 open System.Collections.Generic
 
+open Ionide.LanguageServerProtocol.Server
 open Ionide.LanguageServerProtocol.Types
 open Ionide.LanguageServerProtocol.Types.LspResult
 open Microsoft.CodeAnalysis
-open Microsoft.CodeAnalysis.Text
 open Microsoft.CodeAnalysis.Classification
+open Microsoft.CodeAnalysis.Text
 
 open CSharpLanguageServer.State
 open CSharpLanguageServer.Util
 open CSharpLanguageServer.Conversions
+open CSharpLanguageServer.Types
 
 [<RequireQualifiedAccess>]
 module SemanticTokens =
@@ -138,12 +140,35 @@ module SemanticTokens =
             return Some response |> success
     }
 
+    let private dynamicRegistration (clientCapabilities: ClientCapabilities option) =
+        clientCapabilities
+        |> Option.bind (fun x -> x.TextDocument)
+        |> Option.bind (fun x -> x.SemanticTokens)
+        |> Option.bind (fun x -> x.DynamicRegistration)
+        |> Option.defaultValue false
+
     let provider (clientCapabilities: ClientCapabilities option) : SemanticTokensOptions option =
-        Some { Legend = { TokenTypes = semanticTokenTypes |> Seq.toArray
-                          TokenModifiers = semanticTokenModifiers |> Seq.toArray }
-               Range = Some true
-               Full = true |> First |> Some
-             }
+        match dynamicRegistration clientCapabilities with
+        | true -> None
+        | false ->
+            Some { Legend = { TokenTypes = semanticTokenTypes |> Seq.toArray
+                              TokenModifiers = semanticTokenModifiers |> Seq.toArray }
+                   Range = Some true
+                   Full = true |> First |> Some }
+
+    let registration (clientCapabilities: ClientCapabilities option) : Registration option =
+        match dynamicRegistration clientCapabilities with
+        | false -> None
+        | true ->
+            Some
+                { Id = Guid.NewGuid().ToString()
+                  Method = "textDocument/semanticTokens"
+                  RegisterOptions =
+                      { Legend = { TokenTypes = semanticTokenTypes |> Seq.toArray
+                                   TokenModifiers = semanticTokenModifiers |> Seq.toArray }
+                        Range = Some true
+                        Full = true |> First |> Some
+                        DocumentSelector = Some defaultDocumentSelector } |> serialize |> Some }
 
     // TODO: Everytime the server will re-compute semantic tokens, is it possible to cache the result?
     let handleFull (scope: ServerRequestScope) (p: SemanticTokensParams): AsyncLspResult<SemanticTokens option> =